Knockout 3.5.0 upgrade:
disposeWhenNodeIsRemoved is a computed option, not an extender. I have been doing it wrong in a lot of cases, but the Typescript typings saved the day \o/ and a lot of memory leaks! I did run into knockout/issues/2471, where the disposeWhenNodeIsRemoved has the wrong typing in the Typescript definition. For now I've cast all input values to a bool.